identify_outliers main

This function runs the data modeling and statistical test for the hypothesis that a transcript includes outlier biological replicate.

identify_outliers(
  .data,
  formula = ~1,
  .sample,
  .transcript,
  .abundance,
  .significance,
  .do_check,
  percent_false_positive_genes = 1,
  how_many_negative_controls = 500,
  approximate_posterior_inference = TRUE,
  approximate_posterior_analysis = TRUE,
  draws_after_tail = 10,
  save_generated_quantities = FALSE,
  additional_parameters_to_save = c(),
  cores = detect_cores(),
  pass_fit = FALSE,
  do_check_only_on_detrimental = length(parse_formula(formula)) > 0,
  tol_rel_obj = 0.01,
  just_discovery = FALSE,
  seed = sample(seq_len(length.out = 999999), size = 1),
  adj_prob_theshold_2 = NULL,
  return_fit = FALSE
)

Arguments

.data
A tibble including a transcript name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column

formula
A formula. The sample formula used to perform the differential transcript abundance analysis

.sample
A column name as symbol. The sample identifier

.transcript
A column name as symbol. The transcript identifier

.abundance
A column name as symbol. The transcript abundance (read count)

.significance
A column name as symbol. A column with the Pvalue, or other significance measure (preferred Pvalue over false discovery rate)

.do_check
A column name as symbol. A column with a boolean indicating whether a transcript was identified as differentially abundant

percent_false_positive_genes
A real between 0 and 100. It is the aimed percent of transcript being a false positive. For example, percent_false_positive_genes = 1 provide 1 percent of the calls for outlier containing transcripts that has actually not outliers.

how_many_negative_controls
An integer. How many transcript from the bottom non-significant should be taken for inferring the mean-overdispersion trend.

approximate_posterior_inference
A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes It confers execution time advantage.

approximate_posterior_analysis
A boolean. Whether the calculation of the credible intervals should be done semi-analytically, rather than with pure sampling from the posterior. It confers execution time and memory advantage.

draws_after_tail
An integer. How many draws should on average be after the tail, in a way to inform CI.

save_generated_quantities
A boolean. Used for development and testing purposes

additional_parameters_to_save
A character vector. Used for development and testing purposes

cores
An integer. How many cored to be used with parallel calculations.

pass_fit
A boolean. Used for development and testing purposes

do_check_only_on_detrimental
A boolean. Whether to test only for detrimental outliers (same direction as the fold change). It allows to test for less transcript/sample pairs and therefore higher the probability threshold.

tol_rel_obj
A real. Used for development and testing purposes

just_discovery
A boolean. Used for development and testing purposes

seed
An integer. Used for development and testing purposes

adj_prob_theshold_2
A boolean. Used for development and testing purposes

return_fit
A boolean

Value

A nested tibble tbl with transcript-wise information: sample wise data | plot | ppc samples failed | tot deleterious outliers

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data("counts")

result =
  counts %>%
  dplyr::mutate(  is_significant = ifelse(symbol %in% c("SLC16A12", "CYP1A1", "ART3"), TRUE, FALSE) ) %>%
 ppcseq::identify_outliers(
  formula = ~ Label,
  sample, symbol, value,
  .significance = PValue,
  .do_check  = is_significant,
  percent_false_positive_genes = 1,
  tol_rel_obj = 0.01,
  approximate_posterior_inference =TRUE,
  approximate_posterior_analysis =TRUE,
  how_many_negative_controls = 50,
  cores=1
)

#> No group or design set. Assuming all samples belong to one group.

#> executing do_inference

#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.

#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00055 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.5 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Iteration: 250 / 250 [100%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 0.1].
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100      -452505.296             1.000            1.000
#> Chain 1:    200      -157418.505             1.437            1.875
#> Chain 1:    300       -70222.774             1.372            1.242
#> Chain 1:    400       -35745.989             1.270            1.242
#> Chain 1:    500       -22618.086             1.132            1.000
#> Chain 1:    600       -16658.317             1.003            1.000
#> Chain 1:    700       -13837.206             0.889            0.964
#> Chain 1:    800       -12322.431             0.793            0.964
#> Chain 1:    900       -11506.892             0.713            0.580
#> Chain 1:   1000       -11028.449             0.646            0.580
#> Chain 1:   1100       -10717.003             0.590            0.358   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1200       -10464.457             0.543            0.358   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1300       -10245.313             0.503            0.204   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1400       -10072.091             0.468            0.204
#> Chain 1:   1500        -9894.544             0.438            0.123
#> Chain 1:   1600        -9736.109             0.412            0.123
#> Chain 1:   1700        -9594.897             0.388            0.071
#> Chain 1:   1800        -9446.112             0.368            0.071
#> Chain 1:   1900        -9342.712             0.349            0.043
#> Chain 1:   2000        -9209.518             0.332            0.043
#> Chain 1:   2100        -9088.416             0.317            0.029
#> Chain 1:   2200        -8987.869             0.303            0.029
#> Chain 1:   2300        -8886.049             0.290            0.024
#> Chain 1:   2400        -8791.583             0.279            0.024
#> Chain 1:   2500        -8716.956             0.268            0.021
#> Chain 1:   2600        -8635.481             0.258            0.021
#> Chain 1:   2700        -8552.895             0.249            0.018
#> Chain 1:   2800        -8494.802             0.240            0.018
#> Chain 1:   2900        -8427.819             0.232            0.017
#> Chain 1:   3000        -8362.447             0.225            0.017
#> Chain 1:   3100        -8310.672             0.218            0.016
#> Chain 1:   3200        -8256.785             0.211            0.016
#> Chain 1:   3300        -8208.731             0.205            0.016
#> Chain 1:   3400        -8155.343             0.199            0.016
#> Chain 1:   3500        -8128.213             0.193            0.015
#> Chain 1:   3600        -8080.613             0.188            0.015
#> Chain 1:   3700        -8043.241             0.183            0.014
#> Chain 1:   3800        -8009.696             0.178            0.014
#> Chain 1:   3900        -7978.113             0.174            0.013
#> Chain 1:   4000        -7947.499             0.170            0.013
#> Chain 1:   4100        -7916.283             0.166            0.011
#> Chain 1:   4200        -7889.169             0.162            0.011
#> Chain 1:   4300        -7860.121             0.158            0.011
#> Chain 1:   4400        -7838.884             0.155            0.011
#> Chain 1:   4500        -7818.803             0.151            0.011
#> Chain 1:   4600        -7794.792             0.148            0.011
#> Chain 1:   4700        -7772.819             0.145            0.011
#> Chain 1:   4800        -7750.911             0.142            0.011
#> Chain 1:   4900        -7728.804             0.139            0.010
#> Chain 1:   5000        -7712.898             0.136            0.010
#> Chain 1:   5100        -7701.366             0.116            0.009
#> Chain 1:   5200        -7678.364             0.079            0.009
#> Chain 1:   5300        -7663.574             0.054            0.008
#> Chain 1:   5400        -7655.062             0.035            0.008
#> Chain 1:   5500        -7640.610             0.023            0.007
#> Chain 1:   5600        -7627.211             0.016            0.007
#> Chain 1:   5700        -7620.446             0.012            0.007
#> Chain 1:   5800        -7605.334             0.010            0.006
#> Chain 1:   5900        -7594.515             0.008            0.006
#> Chain 1:   6000        -7586.622             0.008            0.006
#> Chain 1:   6100        -7578.763             0.007            0.005   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

#> Warning: Pareto k diagnostic value is 5.08. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.

#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng

#> Warning: Unknown columns: `Rhat`

#> executing do_inference

#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.

#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000581 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.81 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -12030.210             1.000            1.000
#> Chain 1:    200        -8451.246             0.712            1.000
#> Chain 1:    300        -7581.437             0.513            0.423
#> Chain 1:    400        -7565.731             0.385            0.423
#> Chain 1:    500        -7520.117             0.309            0.115
#> Chain 1:    600        -7510.318             0.258            0.115
#> Chain 1:    700        -7523.554             0.221            0.006
#> Chain 1:    800        -7499.543             0.194            0.006
#> Chain 1:    900        -7498.006             0.173            0.003   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

#> Warning: Pareto k diagnostic value is 2.54. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.

#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng_approximated

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: Unknown columns: `generated quantities`