plot_credible interval for theoretical data distributions

Plot the data along the theoretical data distribution.

Usage,

plot_credible_intervals(.data)

Arguments

.data

The tibble returned by identify_outliers

Value

A tibble with an additional plot column

Examples

library(dplyr)

data("counts")

if(Sys.info()[['sysname']] == "Linux"){
result =
  counts %>%
  dplyr::mutate(  is_significant = ifelse(symbol %in% c("SLC16A12", "CYP1A1", "ART3"), TRUE, FALSE) ) %>%
 ppcseq::identify_outliers(
  formula = ~ Label,
  sample, symbol, value,
  .significance = PValue,
  .do_check  = is_significant,
  percent_false_positive_genes = 1,
  tol_rel_obj = 0.01,
  approximate_posterior_inference =TRUE,
  approximate_posterior_analysis =TRUE,
  how_many_negative_controls = 50,
  cores=1
)

result_plot = result %>% plot_credible_intervals()
}
#> No group or design set. Assuming all samples belong to one group.
#> executing do_inference
#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000521 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.21 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100   -156518854.612             1.000            1.000
#> Chain 1:    200   -535614774.272             0.854            1.000
#> Chain 1:    300   -193453644.893             1.159            1.000
#> Chain 1:    400    -60320697.166             1.421            1.769
#> Chain 1:    500   -276717647.851             1.293            1.000
#> Chain 1:    600     -6280000.684             8.255            1.769
#> Chain 1:    700      -955856.469             7.871            1.769
#> Chain 1:    800  -62139499914559.266             7.012            1.769
#> Chain 1:    900  -1628027603784.084            10.363            1.769
#> Chain 1:   1000  -2455530169.140            75.527            2.207
#> Chain 1:   1100    -83760294.615            71.235            2.207   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1200       -34993.394           264.683            5.570   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1300       -22841.209           244.363            2.207   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1400      -936966.782           226.978            2.207   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1500     -2730460.985           211.890            1.769   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1600       -11252.179           213.751            2.207   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1700        -9627.406           201.187            1.769   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1800        -8593.099           190.017            1.769   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1900        -8137.619           180.019            1.000   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2000        -7770.834           171.020            1.000   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2100        -7591.228           162.878            1.000   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2200        -7536.715           155.475            1.000   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2300        -7522.356           148.715            0.976   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2400        -7531.175           142.518            0.976   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2500        -7521.998           136.818            0.782   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2600        -7523.608           131.556            0.782   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2700        -7522.516           126.683            0.708   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2800        -7527.706           122.159            0.708   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   2900        -7526.060           117.946            0.657   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3000        -7520.666           114.015            0.657   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3100        -7541.812           110.337            0.532   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3200        -7518.740           106.889            0.532   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3300        -7518.885           103.650            0.169   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3400        -7513.435           100.602            0.169   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3500        -7515.823            97.727            0.120   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3600        -7510.898            95.013            0.120   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3700        -7511.633            92.445            0.056   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3800        -7511.017            90.012            0.056   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   3900        -7518.918            87.704            0.047   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4000        -7516.442            85.511            0.047   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4100        -7513.626            83.426            0.024   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4200        -7512.469            81.439            0.024   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4300        -7509.624            79.545            0.007   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4400        -7513.173            77.738            0.007   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   4500        -7520.867            76.010            0.003   MEDIAN ELBO CONVERGED   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 2.36. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng
#> Warning: Unknown columns: `Rhat`
#> executing do_inference
#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000546 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.46 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -12041.890             1.000            1.000
#> Chain 1:    200        -8488.332             0.709            1.000
#> Chain 1:    300        -7587.794             0.512            0.419
#> Chain 1:    400        -7506.376             0.387            0.419
#> Chain 1:    500        -7516.220             0.310            0.119
#> Chain 1:    600        -7495.500             0.259            0.119
#> Chain 1:    700        -7501.266             0.222            0.011
#> Chain 1:    800        -7495.382             0.194            0.011
#> Chain 1:    900        -7497.096             0.173            0.003   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 3.06. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng_approximated
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: Unknown columns: `generated quantities`