plot_credible interval for theoretical data distributions

Plot the data along the theoretical data distribution.

plot_credible_intervals(.data)

Arguments

.data
The tibble returned by identify_outliers

Value

A tibble with an additional plot column

Examples

library(dplyr)

data("counts")

result =
  counts %>%
  dplyr::mutate(  is_significant = ifelse(symbol %in% c("SLC16A12", "CYP1A1", "ART3"), TRUE, FALSE) ) %>%
 ppcseq::identify_outliers(
  formula = ~ Label,
  sample, symbol, value,
  .significance = PValue,
  .do_check  = is_significant,
  percent_false_positive_genes = 1,
  tol_rel_obj = 0.01,
  approximate_posterior_inference =TRUE,
  approximate_posterior_analysis =TRUE,
  how_many_negative_controls = 50,
  cores=1
)

#> No group or design set. Assuming all samples belong to one group.

#> executing do_inference

#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.

#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000518 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.18 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -11981.662             1.000            1.000
#> Chain 1:    200        -8557.859             0.700            1.000
#> Chain 1:    300        -7613.530             0.508            0.400
#> Chain 1:    400        -7553.523             0.383            0.400
#> Chain 1:    500        -7534.003             0.307            0.124
#> Chain 1:    600        -7524.601             0.256            0.124
#> Chain 1:    700        -7519.383             0.220            0.008
#> Chain 1:    800        -7526.333             0.192            0.008
#> Chain 1:    900        -7522.442             0.171            0.003   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

#> Warning: Pareto k diagnostic value is 2.81. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.

#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng

#> Warning: Unknown columns: `Rhat`

#> executing do_inference

#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.

#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Rejecting initial value:
#> Chain 1:   Gradient evaluated at the initial value is not finite.
#> Chain 1:   Stan can't start sampling from this initial value.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.001595 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 15.95 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100      -129015.024             1.000            1.000
#> Chain 1:    200       -42386.895             1.522            2.044
#> Chain 1:    300       -17986.456             1.467            1.357
#> Chain 1:    400       -14186.116             1.167            1.357
#> Chain 1:    500       -10619.154             1.001            1.000
#> Chain 1:    600        -8489.663             0.876            1.000
#> Chain 1:    700        -7750.016             0.764            0.336
#> Chain 1:    800        -7568.890             0.672            0.336
#> Chain 1:    900        -7522.303             0.598            0.268
#> Chain 1:   1000        -7533.607             0.538            0.268
#> Chain 1:   1100        -7522.349             0.489            0.251
#> Chain 1:   1200        -7506.532             0.449            0.251
#> Chain 1:   1300        -7502.359             0.414            0.095
#> Chain 1:   1400        -7505.259             0.385            0.095
#> Chain 1:   1500        -7497.871             0.359            0.024
#> Chain 1:   1600        -7503.176             0.337            0.024
#> Chain 1:   1700        -7500.462             0.317            0.006
#> Chain 1:   1800        -7494.996             0.299            0.006
#> Chain 1:   1900        -7498.758             0.284            0.002   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

#> Warning: Pareto k diagnostic value is 2.79. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.

#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng_approximated

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: first element used of 'length.out' argument

#> Warning: Unknown columns: `generated quantities`

result_plot = result %>% plot_credible_intervals()