plot_credible interval for theoretical data distributions

Plot the data along the theoretical data distribution.

plot_credible_intervals(.data)

Arguments

.data

The tibble returned by identify_outliers

Value

A tibble with an additional plot column

Examples

library(dplyr)

data("counts")

if(Sys.info()[['sysname']] == "Linux"){
result =
  counts %>%
  dplyr::mutate(  is_significant = ifelse(symbol %in% c("SLC16A12", "CYP1A1", "ART3"), TRUE, FALSE) ) %>%
 ppcseq::identify_outliers(
  formula = ~ Label,
  sample, symbol, value,
  .significance = PValue,
  .do_check  = is_significant,
  percent_false_positive_genes = 1,
  tol_rel_obj = 0.01,
  approximate_posterior_inference =TRUE,
  approximate_posterior_analysis =TRUE,
  how_many_negative_controls = 50,
  cores=1
)

result_plot = result %>% plot_credible_intervals()
}
#> No group or design set. Assuming all samples belong to one group.
#> executing do_inference
#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000509 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.09 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -12379.331             1.000            1.000
#> Chain 1:    200        -8612.603             0.719            1.000
#> Chain 1:    300        -7625.216             0.522            0.437
#> Chain 1:    400        -7545.113             0.394            0.437
#> Chain 1:    500        -7526.712             0.316            0.129
#> Chain 1:    600        -7526.704             0.263            0.129
#> Chain 1:    700        -7527.180             0.226            0.011
#> Chain 1:    800        -7522.051             0.198            0.011
#> Chain 1:    900        -7526.694             0.176            0.002   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 2.52. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng
#> Warning: Unknown columns: `Rhat`
#> executing do_inference
#> Storing counts in `nn`, as `n` already present in input
#> ℹ Use `name = "new_name"` to pick a new name.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Rejecting initial value:
#> Chain 1:   Gradient evaluated at the initial value is not finite.
#> Chain 1:   Stan can't start sampling from this initial value.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000454 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.54 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -12374.808             1.000            1.000
#> Chain 1:    200        -8578.377             0.721            1.000
#> Chain 1:    300        -7585.113             0.525            0.443
#> Chain 1:    400        -7515.663             0.396            0.443
#> Chain 1:    500        -7513.298             0.317            0.131
#> Chain 1:    600        -7547.252             0.265            0.131
#> Chain 1:    700        -7508.058             0.228            0.009
#> Chain 1:    800        -7504.691             0.199            0.009
#> Chain 1:    900        -7501.627             0.177            0.005
#> Chain 1:   1000        -7503.691             0.159            0.005
#> Chain 1:   1100        -7508.659             0.145            0.004   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 3.32. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
#> executing add_exposure_rate
#> executing check_if_within_posterior
#> executing save_generated_quantities_in_case
#> executing fit_to_counts_rng_approximated
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: first element used of 'length.out' argument
#> Warning: Unknown columns: `generated quantities`